Restructure derive crate
